pip-requirements.txt
pyyaml
palettable
